% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bd_process_test_s}
\alias{bd_process_test_s}
\title{Bayesian SNCP fit using BD-MCMC}
\usage{
bd_process_test_s(obs_points, mean_mu_alpha, sd_log_alpha, sd_prop_alpha, beta,
  n_it, window_hw, df_iw_prior, df_iw_prop, sigma_prior, lung_data,
  var_mu_alpha, pen_dist, pen_val, n_cent_init, prior_n_cent, max_bd_events,
  max_bd_vt, sigma_smooth)
}
\arguments{
\item{obs_points}{matrix of coordinates for observed points}

\item{mean_mu_alpha}{prior mean for mu_alpha}

\item{sd_log_alpha}{prior std. dev. for log_alphas}

\item{sd_prop_alpha}{std. dev. for random walk proposal for log_alphas}

\item{beta}{prior estimate of homogeneous Poisson intensity function for cluster centers}

\item{n_it}{number of iterations to run MCMC}

\item{window_hw}{half-width of random walk cube for proposing new cluster center locations}

\item{df_iw_prior}{degrees of freedom for inverse-wishart prior on dispersion matrices}

\item{df_iw_prop}{degrees of freedom for inverse-wishart proposal on dispersion matrices}

\item{sigma_prior}{prior for dispersion matrices sigma}

\item{lung_data}{matrix that contains coordinates of all points within 2d slice of lung CT that is lung tissue}

\item{var_mu_alpha}{prior variance on mu_alpha}

\item{pen_dist}{distance for strauss process repulsion}

\item{pen_val}{vaule between 0 and 1 that is the penalty for 2 cluster centers being within pen_dist of each other}

\item{n_cent_init}{number of clusters to start with}

\item{prior_n_cent}{prior value for number of clusters}

\item{max_bd_events}{max events (births + deaths) to allow at each iteration of BD-MCMC}

\item{max_bd_vt}{max ammount of virtual time to spend in BD process at each BD_MCMC iteration}

\item{sigma_smooth}{smoothing bandwidth for BD proposal surface}
}
\value{
Returns a named list posterior samples of model parameters
}
\description{
Run a BD-MCMC chain for a SNCP (mvn dispersion density) on lung ct data
}
\author{
anon
}
