% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundry_total}
\alias{boundry_total}
\title{Boundry pixels of the lung in an example axial slice from a CT scan.}
\format{A data frame with 614 rows and 2 variables:
\describe{
  \item{x}{x coordinate}
  \item{y}{y coordinate}
  ...
}}
\usage{
boundry_total
}
\description{
A dataframe that contains the x and y coordinates of the boundary pixels for a single lung in an axial CT slice
}
\keyword{datasets}
