% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastPdist2}
\alias{fastPdist2}
\title{Fast p-dist2 function}
\usage{
fastPdist2(A, B)
}
\arguments{
\item{A}{matrix of points}

\item{B}{matrix of points}
}
\value{
Returns a matrix with distances
}
\description{
Run p-dist
}
\details{
This is where you write details on the function...

more details....
}
\author{
Brian Vestal
}
