% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lung_data}
\alias{lung_data}
\title{All pixels of lung tissue in an example axial slice from a CT scan.}
\format{
A data frame with 25,473 rows and 2 variables:
\describe{
  \item{x}{x coordinate}
  \item{y}{y coordinate}
  ...
}
}
\usage{
lung_data
}
\description{
A dataframe that contains the x and y coordinates for all of the pixels in a single lung in one axial CT slice
}
\keyword{datasets}
