% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sncp_mcmc_cont_fixed}
\alias{sncp_mcmc_cont_fixed}
\title{Bayesian SNCP fit using BD-MCMC}
\usage{
sncp_mcmc_cont_fixed(
  obs_points,
  centers,
  mean_mu_alpha,
  sd_log_alpha,
  sd_prop_alpha,
  beta,
  n_it,
  window_hw,
  df_iw_prior,
  df_iw_prop,
  sigma_prior,
  xwin,
  ywin,
  var_mu_alpha,
  pen_dist,
  pen_val,
  prior_n_cent,
  max_bd_events,
  max_bd_vt
)
}
\arguments{
\item{obs_points}{matrix of coordinates for observed points}

\item{mean_mu_alpha}{prior mean for mu_alpha}

\item{sd_log_alpha}{prior std. dev. for log_alphas}

\item{sd_prop_alpha}{std. dev. for random walk proposal for log_alphas}

\item{beta}{prior estimate of homogeneous Poisson intensity function for cluster centers}

\item{n_it}{number of iterations to run MCMC}

\item{window_hw}{half-width of random walk cube for proposing new cluster center locations}

\item{df_iw_prior}{degrees of freedom for inverse-wishart prior on dispersion matrices}

\item{df_iw_prop}{degrees of freedom for inverse-wishart proposal on dispersion matrices}

\item{sigma_prior}{prior for dispersion matrices sigma}

\item{xwin}{vector that has the min and max x values of the observation window}

\item{ywin}{vector that has the min and max y values of the observation window}

\item{var_mu_alpha}{prior variance on mu_alpha}

\item{pen_dist}{distance for strauss process repulsion}

\item{pen_val}{vaule between 0 and 1 that is the penalty for 2 cluster centers being within pen_dist of each other}

\item{prior_n_cent}{prior value for number of clusters}

\item{max_bd_events}{max events (births + deaths) to allow at each iteration of BD-MCMC}

\item{max_bd_vt}{max ammount of virtual time to spend in BD process at each BD_MCMC iteration}

\item{n_cent_init}{number of clusters to start with}
}
\value{
Returns a named list containing posterior samples of model parameters with the following elements:
\itemize{
\item log_alpha_sample = list whose elements are vectors containing the estimated log-alpha estimates for the clusters present in that iteration
\item centers_sample = list whose elements matrices containing the estimated center locations for the clusters present in that iteration
\item sigmas_sample = list whose elements are lists containing the estimated covariance matrix estimates for the clusters present in that iteration
\item mu_alpha_sample = vector of mu_alpha samples
\item eta_sample = vector of eta samples,
\item n_centers_sample = vector that contains the number of estimated clusters at each iteration,
\item cumulative_intensity_sample = vector that contains the cumulative intensity estimate for the SNCP at each iteration
}
}
\description{
Run a BD-MCMC chain for a SNCP (mvn dispersion density) on lung ct data using a uniform proposal surface for the BD process
}
\author{
Brian Vestal
}
